next attempt at shellcheck

+++ usr/local/share/bastille/destroy.sh
@@ -178,7 +178,7 @@ destroy_rel() {
             if [ "${FORCE}" = "1" ]; then
                 ## remove cache on force
                 if [ -d "${bastille_cachedir}/${TARGET}" ]; then
-                    rm -rf "${bastille_cachedir}/${TARGET}"
+                    rm -rf "${bastille_cachedir:?}/${TARGET:?}"
                 fi
             fi
             echo
@@ -203,7 +203,7 @@ while [ "$#" -gt 0 ]; do
             FORCE="1"
             shift
             ;;
-        -*|--*)
+        -*)
             error_exit "Unknown Option: \"${1}\""
             usage
             ;;
